Add support for different shell init files
This commit adds support for different shell init files in the installation script. The script now checks if the current shell is Bash or Zsh and sets the appropriate init file accordingly. If the shell is not supported, an error message is displayed. This change allows the library to be installed correctly in different shell environments.